Basic event passing through ring buffer
This patch adds some super simple event passing from kernel to userspace
via a ringbuffer. It also adds bind generation of C types used by the
bpf code to Rust userspace, this in turns allows the types to be
translated to native Rust types and deal with them in a more idiomatic
way.