ROX-30259: Add a basic integration test framework
The tests are based on pytest, since it should be the easiest way to get
them to run on as many systems as possible natively. The only test that
has been implemented is incredibly simple and naive, just create a new
directory, run fact monitoring said directory create a file in there and
expect an event from fact in a gRPC server.

In order to work around a verifier issue, path filtering has been moved
from the kernel to userspace, this is not ideal and we might need to
come back and fix it sometime in the future.

Next step will be to flesh out the tests and integrate them with the
collector ansible scripts to run them on GCP.