feat: Implement an extensible output component
In order to make fact more flexible, an output component has been
implemented which can be extended to run different types of tasks for
outputting the events generated by the eBPF worker. This works by using
a broadcast channel from tokio to forward the events from the eBPF
worker to multiple tasks which will handle things like forwarding via
gRPC, printing to stdout, etc...

The current implementation has an output for gRPC and one printing the
events as JSON to stdout.