Skip emptry dentry when evaluating cgroup path
Currently if faced an empty string in the cgroup path get_cpu_cgroup
bails out without taking into account at which point in processing we
are. It turns out that at least on openshift every cgroup path ends with
an empty segment, meaning that we need to filter them out first.