ROX-30259: Add a basic integration test framework (#1)
* ROX-30259: Add a basic integration test framework

The tests are based on pytest, since it should be the easiest way to get
them to run on as many systems as possible natively. The only test that
has been implemented is incredibly simple and naive, just create a new
directory, run fact monitoring said directory create a file in there and
expect an event from fact in a gRPC server.

In order to work around a verifier issue, path filtering has been moved
from the kernel to userspace, this is not ideal and we might need to
come back and fix it sometime in the future.

Next step will be to flesh out the tests and integrate them with the
collector ansible scripts to run them on GCP.

* Add some AI generated docstrings

* Make the healthcheck optional and use it in the integration tests

* Integration tests don't work on CI yet

* Fix macro formatting and move some general logic to FileActivityService

* Add issue tracker for TODO item

* Rename healthcheck to health_check

* Add integration tests on CI (#4)

* test: Add basis to run integration tests on GCP

* Run integration tests on GCP

* Generate logs and test results and upload them as artifacts

* Add error to test for checking handling of errors on CI

* Revert "Add error to test for checking handling of errors on CI"

This reverts commit 0d99ef6cbaaa9b0c47ed0ac641c402a7b1ff5a30.

* Simplify constants.mk and minor cleanup

* Rename logs fixture to logs_dir