Compare host and task mount namespaces (#16)
is_external_mount should indicate that the task mount namespace is
different from the host mount namespace. The latter could be extracted
from the currently running fact process, assuming it's running in a
privileged mode.

Since in that case the host mount namespace is a property of the fact
process, change the BPF program to return only the namespace and compare
with it in user space.

After this change host_file is set unconditionally, since there is no
is_external_mount in the proble anymore.