Use memory cgroup instead of cpu (#12)
Currently the cpu cgroup is used to extract the container id. It turns
out that on openshift cpu controller is not present in the container
cgroup, only io, memory and pid. This can lead to BPF program not
setting any cgroup value. Switch to the memory cgroup, assuming it's
present more often.

As a side note, there is another approach to find out the container id.
We could extract cgroup_id directly inside the BPF program (there is a
helper for that), and keep a cache of all cgroups found on the system
(via iterating over them at start, then update on demand). When
cgroup_id is received, we should be able to match it with the cgroup
path (cgroup_id == inode of the cgroup path) and use the path to get the
container id. This is similar how bpftrace does it.