ROX-30294: Add prometheus metrics
This patch adds prometheus metrics by leveraging the official
prometheus-client crate. Currently the only metrics are event counting
for the BPF worker and the output components, but extending them should
be relatively straight forward.

A generic EventCounter type has been added in order to somewhat
standardize how we handle event counting by using labels.

In order to expose the metrics, an exporter has been implemented using
hyper (which we were already using for the health check module). This
exporter listens to 0.0.0.0:9001 and is currently hardcoded.