ROX-30294: Add prometheus metrics (#40)
* ROX-30294: Add prometheus metrics

This patch adds prometheus metrics by leveraging the official
prometheus-client crate. Currently the only metrics are event counting
for the BPF worker and the output components, but extending them should
be relatively straight forward.

A generic EventCounter type has been added in order to somewhat
standardize how we handle event counting by using labels.

In order to expose the metrics, an exporter has been implemented using
hyper (which we were already using for the health check module). This
exporter listens to 0.0.0.0:9001 and is currently hardcoded.

* Add kernel metrics to prometheus exporter

* Simplify kernel metric aggregation

* Rename metrics_by_type_t to metrics_by_hook_t

* Improvements to metrics

* Rename `end` label to `error` in file_open hook.
* Rename `dropped` metric in the kernel to `error`
* Add ringbuffer_full metric.
* Add some TODO comments.
* Rename `e` to `event` for consistency in stdout output.

Co-authored-by: Dmitrii Dolgov <9erthalion6@gmail.com>

---------

Co-authored-by: Dmitrii Dolgov <9erthalion6@gmail.com>