ROX-30746: implement layered configuration (#53)
* ROX-30746: implement layered configuration

This patch implements layered configuration by introducing a new
FactConfig type that can be updated with other instances of itself. All
the fields in this struct are optional and initialized to None until a
field is found either in a configuration file or the provided CLI
arguments. Getters are implemented for all fields that will either
return a reference to the held configuration value or use the default
for the field. The object is constructed by calling its new method and
providing a list of files that will be checked in order as provided to
build the configuration, with a final step parsing the CLI arguments and
applying the values extracted from that.

At the time of this patch, configuration files that are parsed are as
follows:
* /etc/stackrox/fact.yml
* /etc/stackrox/fact.yaml
* $(cwd)/fact.yml
* $(cwd)/fact.yaml

Implementation is based on the suggestion from [this
comment](https://github.com/clap-rs/clap/discussions/2763#discussioncomment-1301698)
and the suggested example in the cargo repo
[here](https://github.com/crate-ci/cargo-release/blob/master/src/config.rs)

My initial intention for parsing the configuration files was to use
serde-yaml but, because it has been deprecated, I decided to use
yaml-rust2 and translate from its `Yaml` type to `FactConfig` myself.

* Add unit tests for configuration