chore: Update is_external_mount to be part of the process information (#46)
* chore: Update is_external_mount to be part of the process information

After some discussion, we have arrived to the conclusion that this field
makes more sense as part of the process information, this patch updates
the code to reflect this situation.

Additionally, the bpf code has been updated to directly notify if the
process is running from an external mount namespace by passing the root
namespace inode at load time. With this change, the mnt_namespace is
only read once, so the method in `host_info` is no longer a LazyLock.

* Rename is_external_mount to in_root_mount_ns

* Replace all occurrences of mnt_namespace with mount_ns

This is merely a cosmetic change, it is meant to somewhat simplify our
code by making it more concise and providing coherence in the naming of
variables with the protobuf definitions from the stackrox repo.

* Bump stackrox submodule