Improve integration tests (#58)
* feat(integration-tests): check more process fields

Additionally, the `Process` class can now be constructed based on the
current process or from /proc by providing the pid of a process.

* Add test for ignored directories

* Add a test for multiple files being opened

* Add test for multiple open actions on the same file

* Move executor to FileActivityService class

* Update docstrings

* Add open test for a process other than pytest

* Increase external process sleep

The process needs to exist for the duration of the test, otherwise we
won't be able to create the `Process` object.

* Use multiprocessing module instead of subprocess.Popen

This should be more consistent, as it won't depend on the shell behavior
to perform the action required, while also spawning a new process.

A deprecation warning about forking in a multi-threaded process is
silenced, since we can't really do much about it.