ROX-30254: Track file creation events (#65)
* ROX-30254: Track file creation events

Creation events are issued whenever a file is opened and the
FMODE_CREATED flag is issued. The provided fields match the ones for the
already implemented open event and they are distinguished by a new
`type` flag in the event.

The initial intent for tracking creation events was to use the
inode_create LSM hook. However, the inode_* hooks only provide an inode
and a dentry at best, from this information it is not possible to
reconstruct the path to the file being accessed, since we don't have the
mount point for the dentry. Additionally, an inode can have multiple
files associated with it (hardlinks), so it is also not possible to
reconstruct it from this information.

* Use FileEvent and IsMonitored traits to reduce code duplication

This was suggested by Gemini, not 100% a fan of it, but a bit better
than my original implementation, so I'm keeping it in.