ROX-30744: Log system information at startup (#70)
This change should help narrow down what system an error happens on.

The logged information matches what we do in collector and it includes:
* The running binary version
* The OS distribution
* The kernel version
* The architecture of the CPU

There are a lot of changes around the BPF code, but most of it is just
the subdirectory being turned into a proper rust crate and moving the
bindings into it. This is only done so the build script for the main
crate can focus on getting the git generated version into the built
binary and having the fact-ebpf crate build the bpf code so we don't end
up rebuilding it on every build.