refactor: move log startup logic to lib.rs and simplify uname handling (#75)
Cleanup the main function by moving the log initialization and logging
of system information to lib.rs as functions.

Additionally, the get_kernel_version and get_architecture functions from
the host_info module have been replaced with a SystemInfo type. This
means the underlying LazyLock<HashMap> that was originally used is no
longer needed, the caller can decide what the lifetime of the object
should be and, therefore, we don't need to hang on to that information
for the entire time fact is running.

Finally, the hostname has also been added to the list of information
being logged at start up.