fix: log the correct versions in CI built containers (#76)
After inspecting the test logs from CI it became apparent there are a
few issues. On GHA built containers, the version looks like this:
```
[INFO  2025-09-19T09:37:46Z] (fact/src/lib.rs:53) fact version: 47c1a04967-arm64
```

There are two problems with that version:
* It includes the architecture information as part of the version, which
  it shouldn't do.
* The actual version is the commit SHA instead of the rev-parse digest.

These are fixed by having independent FACT_TAG and FACT_VERSION
variables in the build and checking out the full git history when
determining the version respectively.

On konflux built containers, the version looks like this:
```
[INFO  2025-09-19T09:46:00Z] (fact/src/lib.rs:53) fact version:
```

This is because the `FACT_TAG` argument is only being set on the final
stage of the Containerfile.