ROX-30438: Reimplement kernel side path filtering
Filter events in the kernel by using a LPM (Longest Prefix Match) map
with the configured monitored paths. This type of map has a limit of 256
bytes for filtering, which should be fine for our use case. If a user
complains they are getting too much noise because they are getting
events from a longer path, we can re-implement userspace filtering for
additional checks.