ROX-30714: track cgroup ID to container ID mapping
This change greatly reduces the amount of data sent in the ringbuffer
and the effort the BPF hooks need to retrieve the cgroup of the current
process. In exchange for these benefits, we now need to lookup and keep
track of the cgroups and container IDs that exist on the system
ourselves by iterating over the cgroupsfs.

TODO: Add integration tests with containers.