Add integration test for events generated in containers
Because capturing information for a process running in a container is
not fully possible from `/proc`, some information is passed to the
Process constructor directly to overwrite it. This should be fine
though, because we are using a specific container that should not have
major changes on how it executes the command we are expecting.