ROX-30255: track file deletions (#86)
* ROX-30255: track file deletions

Add a LSM hook to path_unlink to track deletions on files.
Implementation is pretty boilerplaty, the main difference is the hook
provides a path to the directory containing the file and a separate
dentry for the actual file being unlinked, so we need to "paste" the
final path together. Aside from that detail, both the kernel and
userspace implementations are pretty much the same we had for the
previous creation and open events.

Integration tests are also added for the new event type.

* Miscellaneous BPF improvements

* Remove unnecessary checks.
* Add some helpers for path buffer access.
* Add a comment on size of the path_cfg_helper_t buffer size.