refactor: unify hyper endpoints (#90)
Before this patch, the health check and prometheus endpoints both used
hyper and bound separate ports for each. This resulted on quite a bit of
boilerplate being required on both endpoints.

The patch makes it so both endpoints now share the same port and use
some basic routing instead. This should also make it easier in the
future if we want to add any other endpoints for things like
introspection or profiling like we have in collector.

Additionally, disabling all endpoints make it so the hyper task is not
created.