fix(ci): match waiting for images timeout to konflux pipeline timeout
The timeout for the waiting for konflux images job that is used to
trigger the konflux integration tests was set to 20 minutes (with a
comment that mistakenly showed 30), but the konflux pipeline timeout is
set to 40 minutes. Lately, we have had some failures on this job due to
our builds taking slightly longer than 20 minutes, so it makes sense to
have these two timeouts be equal.