ROX-30302: implement heap profiler
This new heap profiler is based on jemalloc and directly handled on the
endpoints module. The implementation is not complete and requires we
replace the allocator we use with jemalloc which has been archived a few
months ago, so we still need to have a talk with the wider team to
decide whether this is something we want to pursue.

For cleanness, the feature is behind a compile time feature, so if we
decide to go with this implementation, we can at least decide when we
want to replace the default allocator.