refactor: simplify event representation
A.K.A cool it with the macros.

The existing events all share the same underlying data, so this is now
represented with a new `BaseEvent` type that is wrapped by all variants
of the `Event` enum.

I have tried to simplify conversion from event_t to `Event` to
`fact_api::FileActivity` as much as possible, but the code is still a
bit of a mess. We could consider using the protobuf generated types
directly, but given all their fields are wrapped in `Option`, I'm fairly
certain that will just make the code even more convoluted.

The event code is also now split, with everything related to process
information living in its own file.