ROX-30836: Implement configuration reloading for endpoints
After implementing this patch it became apparent the effort to fully
implement configuration reloading was too big for a single commit, so
this will be patch 1/N.

A new reloader submodule is implemented for the configuration which will
be in charge of deciding if configuration has been changed, reload it
and notify any relevant tasks that they should handle the new
configuration as required. For testing purposes and simplicity, the
first component to get hot reloading is the exposed hyper endpoints. The
configuration for this module has been split into its own
`EndpointConfig` structure, this makes it possible to notify of changes
to just this subset of fields, which should mean other parts of the code
will be unaffected when only the endpoint configuration is changed.
Ultimately, all configuration will be split in such a way and tasks will
get notified only when relevant configuration is modified.

For simplicity, the reloader will iterate over the list of configuration
files it is meant to monitor every 10 seconds and trigger a reloading
when the modification time of at least one of these has changed. In the
future we could use our own BPF hooks to trigger this condition, but for
now I decided to do a simple as possible implementation.

SIGHUP can also be used to force a reloading of configuration, this is
an established pattern for daemons and also useful for faster testing.