Only parse environment variables and CLI arguments on startup
Environment variables and CLI arguments can only be changed with a
process restart, so there is no point in parsing them every time we hot
reload our configuration. Instead, we can put the parsed configuration
in a `LazyLock` which will be initialized when we first parse
configuration on startup and then re-used when we hot reload.