fix(cli): fix help message starting with `Error: ` (#116)
While testing #110, I realised when running fact with `--help` the usage
message would start with the `Error: ` string, like this:
```
Error: Usage: fact [OPTIONS] [URL]
```

This is due to us incorrectly using the `try_parse` method on our CLI
configuration and propagating the error up. Instead, we can call `parse`
and the application will come to a full stop whenever a wrong argument
is found or the `-h` or `--help` arguments are provided, showing the
correct usage message.

The change also makes it so configuration is parsed before dumping
system information. This is more of a nitpick, having the system
information dumped before the usage message is not a big deal, but I
think this way makes more sense. Additionally, errors parsing
configuration and CLI arguments _should not_ be caused by the actual
system, so I don't think we are losing anything.