feat(hot-reload): reload paths on the fly
This patch makes it so detaching and reattaching programs is no longer
needed for updating the list of monitored paths. This is achieved by
first loading all the new paths into the BPF trie map and then removing
the entries that are not part of the new configuration.

An integration test was added to ensure adding new paths does not remove
existing paths accidentally.