ROX-30630: run BPF bootstrap test in CI
With this change, we will start running our integration tests on GCP
hosted VMs, allowing us to have better visibility into what is the
compatibility of our code with different kernel versions. This is
achieved by running a small bootstrap unit test that load our code into
the kernel, exercising the verifier on it.