ROX-30836: implement hotreloading configuration for outputs (#119)
* ROX-30836: implement hotreloading configuration for outputs

With this patch, it is possible to reload the configuration used for
gRPC communication. Allowing changes to the certificate directory or
output URL to be done without a full restart being required.

The output code has also been slightly refactored, the module now has a
single start function that will spawn a task for gRPC output and another
for JSON to stdout (if this is required at start up or the gRPC output
is not configured). Both the gRPC and stdout code now live as submodules
of the output module. The stdout output is meant mostly for debugging,
so hotreloading was not implemented for it. Lastly, there used to be a
task that would receive `Arc<Event>` messages from the BPF worker,
translate them to the `FileActivity` type and forward them out a
separate channel that ultimately was the gRPC stream, this whole
workflow has now been replaced by a filter_map operation that does the
translating on the gRPC stream directly, leading to more concise code.

* Implement integration test for gRPC hotreloading

* Rename grpc::Client::is_active to is_enabled