ROX-30836: implement hotreloading for monitored paths (#120)
* ROX-30836: implement hotreloading for monitored paths

This patch implements hotreloading for the monitored paths.

In order to achieve this, the BPF worker will need to detach all LSM
hooks before messing with the prefix path map in order to prevent any
problems that may arise of events happening in undefined states. This is
achieved by keeping a list of links for the programs and dropping them,
then re-filling the list once the programs are re-attached. The path
prefix map is cleared by keeping a copy of the list of paths that are
being monitored, then the new list is added in and a copy of it is
stored.

* Add test for monitored paths hotreloading

* Reduce code duplication on BPF programs handling

* Minor cleanup

* feat(hot-reload): reload paths on the fly (#128)

This patch makes it so detaching and reattaching programs is no longer
needed for updating the list of monitored paths. This is achieved by
first loading all the new paths into the BPF trie map and then removing
the entries that are not part of the new configuration.

An integration test was added to ensure adding new paths does not remove
existing paths accidentally.

* Remove duplicated `RINGBUFFER_NAME` const