fix(ci): update grpcio and grpcio-tools to latest release
The VMs we are using for integration tests have been updated and now
have a 3.14 interpreter, the previous version of the grpcio packages we
were using don't have pre-built wheels, so we update to a version that
does.

Additionally, pickle is now failing to serialize the `do_test` functions
in the external process tests, so we move those to the top level.