ROX-30714: use cgroup id for container id resolution
This approach is based on using the cgroup id of processes to resolve
the container ID in userspace. In order to achieve this, a LRU BPF map
is used to keep a mapping of cgroup id to cgroup "path", send the cgroup
id as part of file events and parsing the container id by directly
accessing the map from userspace. The LRU map is filled from userspace
at startup and then kept up to date by a BPF program on the
cgroup_attach_task tracepoint, inserting new cgroup ids as we find them.