Add performance test pipeline (#104)
Add a new pipeline for performance evaluation in three dimensions:

Kernel or user space metric. For user space it's task-clock measured by Perf, or allocated bytes measured by Valgrind. For kernel space it's memlock and run_time / run_cnt as reported by bpftool.

Metric type: CPU utilization or memory allocation.

Workload type:

Ignored workload is simulated by activity (writing) on large number of files, which are not monitored by Fact

Monitored workload is simulated by activity (writing) on a single file, which is monitored by Fact.

Ignored and monitored workload is a mix of both, performed in parallel.

The pipeline is going to be triggered on schedule once a day.

Co-authored-by: Jouko Virtanen <jvirtane@redhat.com>
Co-authored-by: Olivier Valentin <valentio@free.fr>