ROX-30437: track inodes for host path resolution
A HostScanner component is added which will be in charge of going
through the monitored paths on the host and mark them for tracking. It
is also responsible for keeping track of inode to host path mapping so
that events going through fact can have the correct path added to them
when they correspond to a sensitive file on the host. Periodic scans
ensure any missed events will only have a spurious effect.

Kernel side, tracking happens with a hash map that holds the device and
inode numbers for all directories and files that should be tracked.
Initial loading of the map is done by HostScanner, but the kernel will
auto-update it as new files are created. Deleting a file or directory
will also trigger a removal from the map.