ROX-31430: use FIPS mode for gRPC communication
This patch makes it so the TLS library used by tonic changes from ring
to aws-lc-rs and we configure that crate to use FIPS mode. A number of
new dependencies are needed for this crate to be built and are added to
our containerfiles.

A small pre-flight check has been added to ensure we are using FIPS,
this will not fail once the binary is compiled, but if future changes
mess with it the binary will refuse to run and alert us to the
situation.