feat: implement basic pipelining of modules
This patch proposes a simple way of pipelining components. The pattern
is to have each new component in the chain take the receiver end of the
previous block as part of their builder, having the builder return a
tuple with the object and the receiver end of this block, allowing that
receiver to be passed to the next block.

In this new approach there are two special cases:
* The Bpf module starts the chain, so it does not take a receiver.
* The output module is the end of the chain, so it does not give a
  receiver back.