feat: improve handling of path buffers
Since paths can be in non UTF-8 formats, we should strive to use types
that are able to handle this situation when possible. In the case of the
exe_path and filename fields that come from the kernel, handling has
been changed to treat them as binary blobs (same way the kernel d_path
function does) and the length is now being sent to userspace.

With this change, userspace is able to reconstruct a path in a fully
safe way from the path buffer, treating it as a slice and using the
first len - 1 (exclude the null terminator) bytes. This way we are able
to preserve the path all the way to the gRPC event, which requires UTF-8
compliant strings. In a future change we might want to change path
handling in our protobuffs to use byte blobs instead, but this will
probably also require an effort from front end to properly display the
data.